requirements.txt
balethon
google-generativeai